+++ CHANGELOG.rst
@@ -3,5 +3,12 @@
 ============
 
 Version 0.11
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+------------
 -FEATURE- Performance curve regression models can pull data directly from eDNA
+
+Version 0.12
+------------
+- MAJOR- Release of the inference module
+- FEATURE- FailureMode classes, which can determine the current degradation of a failure mode.
+- FEATURE- Evidence classes, which provide supporting evidence for a failure mode
+- FEATURE- ContraryEvidence, which contradict the occurrence of a failure mode